Lesson 14 and 15
Lesson 14: Added a player healthbar
Lesosn 15: Added a gameover screen when play loses.

+++ lesson13/enemy.py
@@ -68,6 +68,6 @@ class Enemy(Sprite):
         
     def fire_bullet(self):
         """ Fire a bullet that will move down. """
-        new_bullet = Bullet(self.settings, self.screen, self, "double")
+        new_bullet = Bullet(self.settings, self.screen, self, "single")
         new_bullet.speed_factor *= -1
         self.bullets_group.add(new_bullet)
